docs/syshealth-finalrelease-doc/final_release_code/sender_side/requirements.txt
digi-xbee==1.3.0
json==2.0.9
mpu6050-raspberrypi==1.0.3.4
smbus2==0.4.1 # or use "apt install python3-smbus"
time==3.9.3
math==3.9.3
# also, "apt install i2c-tools" will be useful when working on the Pi with the mpu